Clean up a bunch of errors, but running into cors issue

+++ packages/rust/conundrum/src/lang/lib/general/pagination/pagination_params.rs
@@ -6,6 +6,6 @@ use typeshare::typeshare;
 #[typeshare]
 #[derive(Debug, Serialize, Deserialize, uniffi::Record)]
 pub struct PaginationParams {
-    pub per_page: u64,
-    pub page: u64,
+    pub per_page: u32,
+    pub page: u32,
 }
