+++ packages/rust/wasm/fluster_wasm/src/lib.rs
@@ -7,8 +7,8 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 pub fn search_conundrum_emojis(query: String, page: usize, per_page: usize) -> Result<JsValue, JsValue> {
     let res = search_emojis(query,
-                            Some(PaginationParams { per_page: per_page as u64,
-                                                    page: page as u64 }));
+                            Some(PaginationParams { per_page: per_page as u32,
+                                                    page: page as u32 }));
     Ok(serde_wasm_bindgen::to_value(&res)?)
 }
